Support for cn-north-1 region
For buckets in cn-north-1 region (and other regions that exist
in "non-standard" AWS partitions), it's not possible to detect
the bucket location from the default endpoint, because cn-north-1
credentials are not valid for the default endpoint.

With this change, the region is first detected using
DefaultAwsRegionProviderChain. If the region doesn't belong to the
standard "aws" partition, we let the AmazonS3 client decide the endpoint.
For regions in the standard partition, or if DefaultAwsRegionProviderChain
can't detect the region, the "old" mechanism of detecting the endpoint is used.